// types.ts
export interface AtlasRegion {
  pageName: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;
  originalWidth: number;
  originalHeight: number;
  rotated: boolean;
  offsetX: number;
  offsetY: number;
  index: number;
}

export type AtlasAssetMap = Map<string, AtlasRegion>;